refactor/ 마이페이지 프로필, 배경 팝업 컴포넌트화
