+++ src/pages/Mypage/MypageEdit.module.css
@@ -1,67 +1,3 @@
-.backgroundEditWrapper {
-  width: 100%;
-  height: 100%;
-  background-color: rgb(192, 191, 191, 0.5);
-  backdrop-filter: blur(4px);
-
-  position: fixed;
-  top: 0;
-  left: 0;
-  z-index: 5;
-
-  transition: ease;
-
-  .backgroundEditContainer {
-    width: 60rem;
-    height: 24rem;
-    background-color: white;
-    border: 1px solid #d9d9d9;
-    border-radius: 10px;
-
-    position: absolute;
-    top: 20%;
-    left: 24%;
-
-    display: flex;
-    flex-direction: column;
-    gap: 10px;
-
-    img {
-      width: 100%;
-      height: 65%;
-    }
-
-    button {
-      width: 5.7rem;
-      height: 2.7rem;
-      border: 1px solid #d9d9d9;
-      border-radius: 5px;
-      background-color: #b99470;
-      color: #ffffff;
-      cursor: pointer;
-      font-size: 1.1rem;
-
-      transition: all 0.6s;
-
-      &:hover {
-        background-color: #ffffff;
-        color: #b99470;
-      }
-    }
-
-    .bgEditCloseBtn {
-      width: 1.3rem;
-      height: 1.9rem;
-      margin-right: 4px;
-      border-radius: 2px;
-      border: none;
-      background-color: transparent;
-
-      transition: background-color 0.4s;
-    }
-  }
-}
-
 .editProfileBtn {
   width: 100%;
   height: 100%;
@@ -505,10 +441,6 @@
 }
 
 @media (max-width: 1400px) {
-  .backgroundEditContainer {
-    width: 45rem !important;
-  }
-
   .editSocialSaveBtn {
     width: 3.9rem !important;
 
@@ -526,10 +458,6 @@
 }
 
 @media (max-width: 1250px) {
-  .backgroundEditContainer {
-    width: 35rem !important;
-  }
-
   .editNamecontainer {
     width: 80%;
     select {
